feat: an user can only CRUD his own meals

+++ src/services/delete-meal.ts
@@ -1,14 +1,24 @@
 import { IMealsRepository } from '@/repositories/meals-repository';
 import { NotFoundError } from './errors/NotFoundError';
+import { UnauthorizedError } from './errors/UnauthorizedError';
+
+interface DeleteMealServiceRequest {
+  userId: string;
+  id: string;
+}
 
 export class DeleteMealService {
   constructor(private mealsRepository: IMealsRepository) {}
 
-  async execute(id: string): Promise<void> {
+  async execute({ id, userId }: DeleteMealServiceRequest): Promise<void> {
     const meal = await this.mealsRepository.findByID(id);
 
     if (!meal) {
       throw new NotFoundError();
+    }
+
+    if (meal.userId !== userId) {
+      throw new UnauthorizedError();
     }
 
     await this.mealsRepository.delete(id);
